Add files via upload
Organized the file system within the project (put scenes in scenes folder, scripts in script folder, etc.) 
Added y sort node to add some depth to characters. 
Changed the collision boxes on the player character and the enemy to be around the feet instead of the whole body so that movement is more fluid.
Changed the enemy movement so that it continuously moves across the screen and changes direction when it hits a wall. (more details in script)